cross-border-system/requirements.txt
pandas==2.0.3
pycryptodome==3.18.0
neo4j==5.12.0
python-dotenv==1.0.0
watchdog==3.0.0
